feat: single route api ready and tested

+++ routers/main.router.ts
@@ -2,6 +2,7 @@ import Database from "../services/db.service";
 import UserInterface from "../models/user.model";
 import { IncomingMessage, ServerResponse } from "http";
 import UserController from "../controllers/user.controller";
+import messages from "../exceptions/api.messages";
 
 export default class Router {
     database: Database<UserInterface>;
@@ -31,45 +32,50 @@ export default class Router {
     proceed(req: IncomingMessage, res: ServerResponse, data: object) {
         const requestUrl = <string>req.url;
 
-        const splitUrl = <string[]>requestUrl.split("/");
+        let splitUrl = <string[]>requestUrl.split("/");
         const method = req.method;
+
+        splitUrl = splitUrl.filter((value: string) => value.length > 0);
 
         let result;
 
         if (
+            splitUrl.length == 3 &&
             splitUrl[0] === "api" &&
             splitUrl[1] === "users" &&
-            splitUrl[2] &&
             method === "GET"
         ) {
             result = this.controller.handleGet(splitUrl[2]);
         } else if (
+            splitUrl.length == 2 &&
             splitUrl[0] === "api" &&
             splitUrl[1] === "users" &&
             method === "GET"
         ) {
             result = this.controller.handleGetAll();
         } else if (
+            splitUrl.length == 2 &&
             splitUrl[0] === "api" &&
             splitUrl[1] === "users" &&
             method === "POST"
         ) {
             result = this.controller.handleAdd(data);
         } else if (
+            splitUrl.length == 3 &&
             splitUrl[0] === "api" &&
             splitUrl[1] === "users" &&
-            splitUrl[2] &&
             method === "PUT"
         ) {
             result = this.controller.handleUpdate(splitUrl[2], data);
         } else if (
+            splitUrl.length == 3 &&
             splitUrl[0] === "api" &&
             splitUrl[1] === "users" &&
             method === "DELETE"
         ) {
             result = this.controller.handleDelete(splitUrl[2]);
         } else {
-            result = { status: 404, message: "Given endpoint doesn't exist." };
+            result = messages.endpointNotFound;
         }
 
         res.setHeader("Content-Type", "application/json");
